Fix up external lkpi, drm, radeon, and amdgpu to build on powerpc (#104)

+++ drivers/gpu/drm/linux_fb.c
@@ -166,7 +166,7 @@ fb_ioctl(struct cdev *dev, u_long cmd, caddr_t data, int fflag,
 
 	case FBIO_BLANK:	/* blank display */
 		if (info->fbops->fb_blank != NULL)
-			error = info->fbops->fb_blank((int)data, info);
+			error = info->fbops->fb_blank((long)data, info);
 		break;
 
 	default:
